+++ src/index.ts
@@ -28,8 +28,8 @@ import SkeletonDirective from './directives/Skeleton/Skeleton'
 import TooltipDirective from './directives/Tooltip/Tooltip'
 
 /* Plugins */
-import Breakpoints from './plugins/Breakpoints/Breakpoints'
-import ToastPlugin from './plugins/Toast/Toast'
+import BreakpointsPlugin from './plugins/Breakpoints'
+import ToastPlugin from './plugins/Toast'
 
 import './styles/main.scss'
 import { App } from 'vue'
@@ -66,7 +66,7 @@ const directives = {
 }
 
 const plugins = {
-  breakpoints: Breakpoints,
+  breakpoints: BreakpointsPlugin,
   toast: ToastPlugin
 }
 
@@ -94,6 +94,8 @@ function install(app: App) {
 }
 
 export default { install }
+
+export { showToast } from './plugins/Toast'
 
 export { default as Avatar } from './components/Avatar'
 export { default as Button } from './components/Button'